feat: deprecate server-data-api (PL-000)
