+++ runtime/lib/DataAPI/index.ts
@@ -1,4 +1,4 @@
 export { default as CreatorDataApi } from './creatorDataAPI';
 export { default as LocalDataApi } from './localDataAPI';
-export { default as ServerDataApi } from './serverDataAPI';
+export { default as ServerDataApi } from './mongoDataAPI';
 export { DataAPI } from './types';
