+++ runtime/lib/DataAPI/types.ts
@@ -9,15 +9,11 @@ export interface DataAPI<
   V extends BaseModels.Version.Model<any> = BaseModels.Version.Model<BaseModels.Version.PlatformData>,
   PJ extends BaseModels.Project.Model<any, any> = BaseModels.Project.Model<AnyRecord, AnyRecord>
 > {
-  init(): Promise<void>;
-
   fetchDisplayById(displayId: number): Promise<null | Display>;
 
   getProgram(programID: string): Promise<P>;
 
   getVersion(versionID: string): Promise<V>;
-
-  unhashVersionID(versionID: string): Promise<string>;
 
   getProject(projectID: string): Promise<PJ>;
